+++ src/main/java/com/him/fpjt/him_backend/auth/controller/AuthController.java
@@ -1,10 +1,15 @@
-package com.him.fpjt.him_backend.user.controller;
+package com.him.fpjt.him_backend.auth.controller;
 
-import com.him.fpjt.him_backend.user.service.AuthService;
+import com.him.fpjt.him_backend.auth.dto.VerificationCodeDto;
+import com.him.fpjt.him_backend.auth.service.AuthService;
+import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
