--- src/main/java/com/him/fpjt/him_backend/user/service/AuthService.java
@@ -1,5 +1,0 @@
-package com.him.fpjt.him_backend.user.service;
-
-public interface AuthService {
-     boolean checkDuplicatedNickname(String nickname);
-}
